Refactor: split checkpoint basis facts
