+++ src/domain/services/optic/CheckpointBasisFactBase.ts
@@ -1,0 +1,16 @@
+import type {
+  CheckpointBasisFactShardFamily,
+  CheckpointBasisFactTransport,
+} from './CheckpointBasisFactTypes.ts';
+
+export abstract class CheckpointBasisFact {
+  abstract readonly kind: CheckpointBasisFactTransport['kind'];
+
+  abstract shardFamily(): CheckpointBasisFactShardFamily;
+
+  abstract shardPath(): string;
+
+  abstract sortKey(): string;
+
+  abstract toTransport(): CheckpointBasisFactTransport;
+}
